[FEAT] Add password visibility toggle in LoginScreen
Added a feature to show and hide password in LoginScreen by importing and using Ionicons. This enhances user experience when logging in, providing users the option to verify their entered password. The implementation involves adding a new state for password visibility and a new toggle icon button within the password input field.